freetrade/pyproject.toml
[tool.poetry]
name = "freetrade"
version = "0.1.0"
description = ""
authors = ["James Walden <james.m.walden@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
google-cloud-storage = "^2.18.2"
pydantic = "^2.9.2"
email-validator = "^2.2.0"
pandas = "^2.2.3"
pydantic-extra-types = "^2.9.0"
pytest = "^8.3.3"
pytest-mock = "^3.14.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
